organizando paas e criando projeto react

+++ .docker/entrypoint.sh
@@ -1,8 +1,17 @@
 #!/bin/bash
 
-#On error no such file entrypoint.sh, execute in terminal - dos2unix .docker\entrypoint.sh
-cp .env.example .env
-cp .env.testing.example .env.testing
+## Frontend
+cd /var/www/frontend && npm install && cd .. 
+
+## Backend
+cd backend
+if [ ! -f ".env" ]; then
+    cp .env.example .env
+fi
+
+if [ ! -f ".env.testing" ]; then
+    cp .env.testing.example .env.testing
+fi
 chown -R www-data:www-data .
 composer install
 php artisan key:generate
